requirements.txt
instabot==0.117.0
Pillow==10.0.1
Requests==2.31.0
